code-gen: move code-generation into Move.thy
Adopted the code-generation mechanism developed in the separate
theory file to the "real" ancestor relation used in the Move.thy
theory.

Tidied up (and Isar'd) the proof of correctness.

Demonstrated that it is more efficient with an example of ancestry
testing in a degenerate tree with a single path of length 26.