code-gen: efficient versions of do-op, undo-op, etc.
Efficient versions of do-op, undo-op, redo-op, interp-op, and
interp-ops provided in terms of hash-maps.  Requires more work than
the ancestry case, as the output of all of these functions is also a
hash-map (not e.g. a bool, as in the case of ancestor).  Also, will
need to show that the main correctness theorem that we have established
for the commutation of interp-ops also holds for these efficient
versions.