evaluation: profiling of remote operation application
Result: 88% of the time is spent in hm_update; 85% spent in
rbt_insert_with_key (rebalancing of red-black tree?). Doesn't look like
it will be easy to speed this up.

Only 8% of the time is spent in hm_lookup, in contrast to applying local
operations, for which hm_lookup is the dominant cost (68%). But since
local operations are vastly faster anyway, hm_lookup doesn't really
matter.